+++ store/store.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/hashicorp/raft"
 	"github.com/hashicorp/raft-boltdb"
-	sql "github.com/rqlite/rqlite/db"
+	sdb "github.com/rqlite/rqlite/db"
 )
 
 var (
@@ -115,8 +115,6 @@ const (
 type Store struct {
 	raftDir string
 
-	mu sync.RWMutex // Sync access between queries and snapshots.
-
 	raft    *raft.Raft // The consensus mechanism.
 	ln      Listener
 	raftTn  *raft.NetworkTransport
@@ -124,10 +122,13 @@ type Store struct {
 	raftLog *raftboltdb.BoltStore // Persisent log store.
 	dbConf  *DBConfig             // SQLite database config.
 	dbPath  string                // Path to underlying SQLite file, if not in-memory.
-	db      *sql.DB               // The underlying SQLite store.
+	db      *sdb.DB               // The underlying SQLite database.
+	dbConn  *sdb.Conn             // Default connection to underlying SQLite database.
 
 	metaMu sync.RWMutex
 	meta   map[string]map[string]string
+
+	restoreMu sync.RWMutex // Restore needs exclusive access to database.
 
 	logger *log.Logger
 
@@ -175,12 +176,17 @@ func (s *Store) Open(enableSingle bool) error {
 		return err
 	}
 
-	// Open underlying database.
-	db, err := s.open()
+	// Create underlying database.
+	if err := s.createDatabase(); err != nil {
+		return err
+	}
+
+	// Get default connection to database.
+	conn, err := s.db.Connect()
 	if err != nil {
 		return err
 	}
-	s.db = db
+	s.dbConn = conn
 
 	// Is this a brand new node?
 	newNode := !pathExists(filepath.Join(s.raftDir, "raft.db"))
@@ -234,7 +240,7 @@ func (s *Store) Open(enableSingle bool) error {
 
 // Close closes the store. If wait is true, waits for a graceful shutdown.
 func (s *Store) Close(wait bool) error {
-	if err := s.db.Close(); err != nil {
+	if err := s.dbConn.Close(); err != nil {
 		return err
 	}
 	f := s.raft.Shutdown()
@@ -382,7 +388,7 @@ func (s *Store) WaitForAppliedIndex(idx uint64, timeout time.Duration) error {
 
 // Stats returns stats for the store.
 func (s *Store) Stats() (map[string]interface{}, error) {
-	fkEnabled, err := s.db.FKConstraints()
+	fkEnabled, err := s.dbConn.FKConstraints()
 	if err != nil {
 		return nil, err
 	}
@@ -390,7 +396,7 @@ func (s *Store) Stats() (map[string]interface{}, error) {
 	dbStatus := map[string]interface{}{
 		"dns":            s.dbConf.DSN,
 		"fk_constraints": enabledFromBool(fkEnabled),
-		"version":        sql.DBVersion,
+		"version":        sdb.DBVersion,
 	}
 	if !s.dbConf.Memory {
 		dbStatus["path"] = s.dbPath
@@ -433,13 +439,17 @@ func (s *Store) Stats() (map[string]interface{}, error) {
 }
 
 // Execute executes queries that return no rows, but do modify the database.
-func (s *Store) Execute(ex *ExecuteRequest) ([]*sql.Result, error) {
+func (s *Store) Execute(ex *ExecuteRequest) ([]*sdb.Result, error) {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
 	return s.execute(ex)
 }
 
 // ExecuteOrAbort executes the requests, but aborts any active transaction
 // on the underlying database in the case of any error.
-func (s *Store) ExecuteOrAbort(ex *ExecuteRequest) (results []*sql.Result, retErr error) {
+func (s *Store) ExecuteOrAbort(ex *ExecuteRequest) (results []*sdb.Result, retErr error) {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
 	defer func() {
 		var errored bool
 		for i := range results {
@@ -449,7 +459,7 @@ func (s *Store) ExecuteOrAbort(ex *ExecuteRequest) (results []*sql.Result, retEr
 			}
 		}
 		if retErr != nil || errored {
-			if err := s.db.AbortTransaction(); err != nil {
+			if err := s.dbConn.AbortTransaction(); err != nil {
 				s.logger.Printf("WARNING: failed to abort transaction: %s", err.Error())
 			}
 		}
@@ -457,7 +467,7 @@ func (s *Store) ExecuteOrAbort(ex *ExecuteRequest) (results []*sql.Result, retEr
 	return s.execute(ex)
 }
 
-func (s *Store) execute(ex *ExecuteRequest) ([]*sql.Result, error) {
+func (s *Store) execute(ex *ExecuteRequest) ([]*sdb.Result, error) {
 	d := &databaseSub{
 		Tx:      ex.Tx,
 		Queries: ex.Queries,
@@ -490,30 +500,23 @@ func (s *Store) execute(ex *ExecuteRequest) ([]*sql.Result, error) {
 // level equivalent to "weak". Otherwise no guarantees are made about the
 // read consistency level.
 func (s *Store) Backup(leader bool, fmt BackupFormat) ([]byte, error) {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
+
 	if leader && s.raft.State() != raft.Leader {
 		return nil, ErrNotLeader
 	}
 
 	var b []byte
+	var err error
 	if fmt == BackupBinary {
-		f, err := ioutil.TempFile("", "rqlite-bak-")
-		if err != nil {
-			return nil, err
-		}
-		f.Close()
-		defer os.Remove(f.Name())
-
-		if err := s.db.Backup(f.Name()); err != nil {
-			return nil, err
-		}
-
-		b, err = ioutil.ReadFile(f.Name())
+		b, err = s.database(leader)
 		if err != nil {
 			return nil, err
 		}
 	} else if fmt == BackupSQL {
 		buf := bytes.NewBuffer(nil)
-		if err := s.db.Dump(buf); err != nil {
+		if err := s.dbConn.Dump(buf); err != nil {
 			return nil, err
 		}
 		b = buf.Bytes()
@@ -525,10 +528,9 @@ func (s *Store) Backup(leader bool, fmt BackupFormat) ([]byte, error) {
 }
 
 // Query executes queries that return rows, and do not modify the database.
-func (s *Store) Query(qr *QueryRequest) ([]*sql.Rows, error) {
-	// Allow concurrent queries.
-	s.mu.RLock()
-	defer s.mu.RUnlock()
+func (s *Store) Query(qr *QueryRequest) ([]*sdb.Rows, error) {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
 
 	if qr.Lvl == Strong {
 		d := &databaseSub{
@@ -561,7 +563,7 @@ func (s *Store) Query(qr *QueryRequest) ([]*sql.Rows, error) {
 		return nil, ErrNotLeader
 	}
 
-	r, err := s.db.Query(qr.Queries, qr.Tx, qr.Timings)
+	r, err := s.dbConn.Query(qr.Queries, qr.Tx, qr.Timings)
 	return r, err
 }
 
@@ -687,28 +689,29 @@ func (s *Store) setMetadata(id string, md map[string]string) error {
 	return nil
 }
 
-// open opens the the in-memory or file-based database.
-func (s *Store) open() (*sql.DB, error) {
-	var db *sql.DB
+// createDatabase creates the the in-memory or file-based database.
+func (s *Store) createDatabase() error {
+	var db *sdb.DB
 	var err error
 	if !s.dbConf.Memory {
 		// as it will be rebuilt from (possibly) a snapshot and committed log entries.
 		if err := os.Remove(s.dbPath); err != nil && !os.IsNotExist(err) {
-			return nil, err
+			return err
 		}
-		db, err = sql.OpenWithDSN(s.dbPath, s.dbConf.DSN)
+		db, err = sdb.New(s.dbPath, s.dbConf.DSN, false)
 		if err != nil {
-			return nil, err
+			return err
 		}
 		s.logger.Println("SQLite database opened at", s.dbPath)
 	} else {
-		db, err = sql.OpenInMemoryWithDSN(s.dbConf.DSN)
+		db, err = sdb.New(s.dbPath, s.dbConf.DSN, true)
 		if err != nil {
-			return nil, err
+			return err
 		}
 		s.logger.Println("SQLite in-memory database opened")
 	}
-	return db, nil
+	s.db = db
+	return nil
 }
 
 // remove removes the node, with the given ID, from the cluster.
@@ -757,12 +760,12 @@ func (s *Store) raftConfig() *raft.Config {
 }
 
 type fsmExecuteResponse struct {
-	results []*sql.Result
+	results []*sdb.Result
 	error   error
 }
 
 type fsmQueryResponse struct {
-	rows  []*sql.Rows
+	rows  []*sdb.Rows
 	error error
 }
 
@@ -772,6 +775,9 @@ type fsmGenericResponse struct {
 
 // Apply applies a Raft log entry to the database.
 func (s *Store) Apply(l *raft.Log) interface{} {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
+
 	var c command
 	if err := json.Unmarshal(l.Data, &c); err != nil {
 		panic(fmt.Sprintf("failed to unmarshal cluster command: %s", err.Error()))
@@ -784,10 +790,10 @@ func (s *Store) Apply(l *raft.Log) interface{} {
 			return &fsmGenericResponse{error: err}
 		}
 		if c.Typ == execute {
-			r, err := s.db.Execute(d.Queries, d.Tx, d.Timings)
+			r, err := s.dbConn.Execute(d.Queries, d.Tx, d.Timings)
 			return &fsmExecuteResponse{results: r, error: err}
 		}
-		r, err := s.db.Query(d.Queries, d.Tx, d.Timings)
+		r, err := s.dbConn.Query(d.Queries, d.Tx, d.Timings)
 		return &fsmQueryResponse{rows: r, error: err}
 	case metadataSet:
 		var d metadataSetSub
@@ -821,47 +827,50 @@ func (s *Store) Apply(l *raft.Log) interface{} {
 	}
 }
 
-// Database returns a copy of the underlying database. The caller should
-// ensure that no transaction is taking place during this call, or an error may
-// be returned. If leader is true, this operation is performed with a read
-// consistency level equivalent to "weak". Otherwise no guarantees are made
-// about the read consistency level.
-//
-// http://sqlite.org/howtocorrupt.html states it is safe to do this
-// as long as no transaction is in progress.
-func (s *Store) Database(leader bool) ([]byte, error) {
+// Database returns a byte slice containing a copy of contents of the
+// underlying SQLite file.
+func (s *Store) database(leader bool) ([]byte, error) {
 	if leader && s.raft.State() != raft.Leader {
 		return nil, ErrNotLeader
 	}
-
-	// Ensure only one snapshot can take place at once, and block all queries.
-	s.mu.Lock()
-	defer s.mu.Unlock()
 
 	f, err := ioutil.TempFile("", "rqlilte-snap-")
 	if err != nil {
 		return nil, err
 	}
-	f.Close()
-	defer os.Remove(f.Name())
+	if err := f.Close(); err != nil {
+		return nil, err
+	}
+	os.Remove(f.Name())
+	db, err := sdb.New(f.Name(), "", false)
+	if err != nil {
+		return nil, err
+	}
+	conn, err := db.Connect()
+	if err != nil {
+		return nil, err
+	}
 
-	if err := s.db.Backup(f.Name()); err != nil {
+	if err := s.dbConn.Backup(conn); err != nil {
+		return nil, err
+	}
+	if err := conn.Close(); err != nil {
 		return nil, err
 	}
 
 	return ioutil.ReadFile(f.Name())
 }
 
-// Snapshot returns a snapshot of the database. The caller must ensure that
-// no transaction is taking place during this call. Hashicorp Raft guarantees
-// that this function will not be called concurrently with Apply.
-//
-// http://sqlite.org/howtocorrupt.html states it is safe to do this
-// as long as no transaction is in progress.
+// Snapshot returns a snapshot of the store. The caller must ensure that
+// no Raft transaction is taking place during this call. Hashicorp Raft
+// guarantees that this function will not be called concurrently with Apply.
 func (s *Store) Snapshot() (raft.FSMSnapshot, error) {
+	s.restoreMu.RLock()
+	defer s.restoreMu.RUnlock()
+
 	fsm := &fsmSnapshot{}
 	var err error
-	fsm.database, err = s.Database(false)
+	fsm.database, err = s.database(false)
 	if err != nil {
 		s.logger.Printf("failed to read database for snapshot: %s", err.Error())
 		return nil, err
@@ -879,9 +888,8 @@ func (s *Store) Snapshot() (raft.FSMSnapshot, error) {
 
 // Restore restores the node to a previous state.
 func (s *Store) Restore(rc io.ReadCloser) error {
-	if err := s.db.Close(); err != nil {
-		return err
-	}
+	s.restoreMu.Lock()
+	defer s.restoreMu.Unlock()
 
 	// Get size of database.
 	var sz uint64
@@ -895,39 +903,32 @@ func (s *Store) Restore(rc io.ReadCloser) error {
 		return err
 	}
 
-	var db *sql.DB
-	var err error
-	if !s.dbConf.Memory {
-		// Write snapshot over any existing database file.
-		if err := ioutil.WriteFile(s.dbPath, database, 0660); err != nil {
-			return err
-		}
-
-		// Re-open it.
-		db, err = sql.OpenWithDSN(s.dbPath, s.dbConf.DSN)
-		if err != nil {
-			return err
-		}
-	} else {
-		// In memory. Copy to temporary file, and then load memory from file.
-		f, err := ioutil.TempFile("", "rqlilte-snap-")
-		if err != nil {
-			return err
-		}
-		f.Close()
-		defer os.Remove(f.Name())
-
-		if err := ioutil.WriteFile(f.Name(), database, 0660); err != nil {
-			return err
-		}
-
-		// Load an in-memory database from the snapshot now on disk.
-		db, err = sql.LoadInMemoryWithDSN(f.Name(), s.dbConf.DSN)
-		if err != nil {
-			return err
-		}
+	// Create temp file and write incoming database to there.
+	temp, err := tempfile()
+	if err != nil {
+		return err
 	}
-	s.db = db
+	defer os.Remove(temp.Name())
+	defer temp.Close()
+
+	if _, err := temp.Write(database); err != nil {
+		return err
+	}
+
+	// Create new database from file, connect, and load
+	// existing database from that.
+	db, err := sdb.New(temp.Name(), "", false)
+	if err != nil {
+		return err
+	}
+	conn, err := db.Connect()
+	if err != nil {
+		return err
+	}
+	defer conn.Close()
+	if err := s.dbConn.Load(conn); err != nil {
+		return err
+	}
 
 	// Read remaining bytes, and set to cluster meta.
 	b, err := ioutil.ReadAll(rc)
@@ -943,6 +944,7 @@ func (s *Store) Restore(rc io.ReadCloser) error {
 	if err != nil {
 		return err
 	}
+
 	stats.Add(numRestores, 1)
 	return nil
 }
@@ -1015,4 +1017,13 @@ func pathExists(p string) bool {
 		return false
 	}
 	return true
+}
+
+// tempfile returns a temporary file for use
+func tempfile() (*os.File, error) {
+	f, err := ioutil.TempFile("", "rqlilte-snap-")
+	if err != nil {
+		return nil, err
+	}
+	return f, nil
 }