+++ CHANGELOG.md
@@ -19,6 +19,7 @@ _To upgrade from an earlier version to this release you should backup the leader
 - [PR #452](https://github.com/rqlite/rqlite/pull/452): Move to Go 1.10.
 - [PR #451](https://github.com/rqlite/rqlite/pull/451), [PR #453](https://github.com/rqlite/rqlite/pull/453): Support dumping database in SQL text format. Fixes [issue #393](https://github.com/rqlite/rqlite/issues/369).
 - [PR #455](https://github.com/rqlite/rqlite/pull/455): Add dump-as-SQL-text command to CLI.
+- [PR #457](https://github.com/rqlite/rqlite/pull/457): Implement new connection-oriented database layer.
 
 ## 4.3.0 (March 18th 2018)
 - [PR #397](https://github.com/rqlite/rqlite/pull/397), [PR #399](https://github.com/rqlite/rqlite/pull/399): Support hashed passwords. Fixes [issue #395](https://github.com/rqlite/rqlite/issues/395). Thanks @sum12.